Store textures in a Weak
When a texture is no longer referenced anymore, it'll be deallocated. We still get the optimization of not needing to load the same texture multiple times, however when loading a texture that's not used anywhere else it'll immediately be deallocated. This is useful for graphic pickers, which shouldn't keep textures loaded